refactor(header-chain): thin planner facade and break planner↔invariants cycle
One-way deps and a thin planner orchestrator; no behavior change.

- Move PlanCandidate / TransitionPlan into planner/plan.rs and re-export
  from the facade / transition::mod so invariants no longer depends on the
  orchestration module for the product type.
- Extract phase-2 glue into planner/replay.rs (BoundRequest,
  bind_replay_and_freshness).
- Keep derive_transition_plan / phase comments in planner.rs as the only
  high-level pipeline narrative.
- Rename planner/evidence.rs → planner/violations.rs (violation taxonomy);
  external re-exports via transition/mod.rs stay identical.
- Extend the overlay include_str! guard to planner.rs, write_set.rs, and
  replay.rs after the split.

Verified: cargo test -p zakura-header-chain --lib
