+++ crates/zakura-header-chain/src/transition/planner.rs
@@ -2,135 +2,40 @@
 
 mod admission;
 mod event_effects;
-mod evidence;
+pub(crate) mod plan;
 mod projected_state;
+mod replay;
 mod settlement;
+mod violations;
 mod write_set;
 
 #[cfg(test)]
 mod tests;
 
-use std::sync::Arc;
-
 use thiserror::Error;
 
-use crate::graph::GraphDelta;
 use crate::{
-    ChangeSet, CounterExhausted, EngineLimits, EngineSnapshot, Frontier, GraphError,
-    HeaderChainEngine, StoreError, TransitionContext, TransitionDomain, TransitionEffect,
+    CounterExhausted, EngineSnapshot, GraphError, HeaderChainEngine, StoreError, TransitionContext,
     TransitionInput,
 };
 
-pub use evidence::{
+pub use violations::{
     AuxiliaryViolation, BodyViolation, FinalityViolation, HeaderPathKind, HeaderPathProblem,
     HeaderValidationCheck, HeaderValidationSource, HeaderViolation, InvalidTransitionEvidence,
     LimitViolation, OperatorViolation, PlannerCoherenceViolation, ProjectionKind,
 };
 
-use admission::{
-    authenticate_and_admit, rebase_header_insertion, validate_body_owner,
-    validate_header_sync_owner, AdmittedRequest, HeaderInsertionRebase,
-};
+pub(crate) use plan::{PlanCandidate, TransitionPlan};
+
+use admission::authenticate_and_admit;
 use event_effects::{apply_event_evidence, migrated_pin_refuted, ApplyEventContext};
 use projected_state::ProjectedTransitionState;
+use replay::bind_replay_and_freshness;
 use settlement::{
     apply_migrated_pin_alarm, derive_finality_and_retention, FinalityRetentionOutcome,
     SettlementInputs,
 };
 use write_set::{derive_plan, invariant_pins, no_change, resource_stalled, DerivePlanInputs};
-
-/// A complete projected write set awaiting independent invariant verification.
-#[derive(Clone, Debug)]
-pub(crate) struct PlanCandidate {
-    pub(super) snapshot_before_commit: EngineSnapshot,
-    pub(super) change_set: ChangeSet,
-    pub(super) graph_delta: GraphDelta,
-    pub(super) domain: TransitionDomain,
-    pub(super) effect: TransitionEffect,
-    pub(super) trust_pins: Arc<[Frontier]>,
-    pub(super) limits: EngineLimits,
-}
-
-impl PlanCandidate {
-    /// Return the opaque graph transition paired with this candidate.
-    pub(super) const fn graph_delta(&self) -> &GraphDelta {
-        &self.graph_delta
-    }
-
-    /// Return the orthogonal transition effects.
-    pub(super) const fn effect(&self) -> TransitionEffect {
-        self.effect
-    }
-}
-
-/// A candidate accepted by the independent transition invariant verifier.
-///
-/// Production callers consume [`crate::EngineTransition`]; this verified type
-/// stays crate-private so adapters cannot construct unchecked transitions.
-#[derive(Clone, Debug)]
-pub(crate) struct TransitionPlan {
-    candidate: PlanCandidate,
-}
-
-impl TransitionPlan {
-    /// Wrap a candidate that has already passed independent invariant verification.
-    fn from_verified(candidate: PlanCandidate) -> Self {
-        Self { candidate }
-    }
-
-    /// Borrow the verified inner candidate (tests and fuzzing only).
-    #[cfg(test)]
-    pub(super) const fn candidate(&self) -> &PlanCandidate {
-        &self.candidate
-    }
-
-    /// Return the atomic write set for the state adapter.
-    pub const fn change_set(&self) -> &ChangeSet {
-        &self.candidate.change_set
-    }
-
-    /// Return the snapshot before commit.
-    pub const fn snapshot_before_commit(&self) -> &EngineSnapshot {
-        &self.candidate.snapshot_before_commit
-    }
-
-    /// Return the submitted transition domain.
-    pub const fn domain(&self) -> TransitionDomain {
-        self.candidate.domain
-    }
-
-    /// Return the orthogonal transition effects.
-    pub const fn effect(&self) -> TransitionEffect {
-        self.candidate.effect
-    }
-
-    /// Return true when the evidence was valid but changed no durable fact.
-    pub fn is_no_change(&self) -> bool {
-        self.candidate.snapshot_before_commit.state_version
-            == self.candidate.change_set.metadata.state_version
-    }
-
-    /// Return the opaque graph transition that matches the durable write set.
-    pub(crate) const fn graph_delta(&self) -> &GraphDelta {
-        &self.candidate.graph_delta
-    }
-}
-
-#[cfg(test)]
-impl std::ops::Deref for TransitionPlan {
-    type Target = PlanCandidate;
-
-    fn deref(&self) -> &Self::Target {
-        &self.candidate
-    }
-}
-
-#[cfg(test)]
-impl std::ops::DerefMut for TransitionPlan {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.candidate
-    }
-}
 
 /// Typed failure that the planner produces before it attempts a durable mutation.
 #[derive(Clone, Debug, Eq, Error, PartialEq)]
@@ -278,84 +183,12 @@ fn derive_plan_candidate(
     )
 }
 
-/// Admitted event after replay and freshness checks (or a short-circuit no-change).
-struct BoundRequest {
-    event: crate::TransitionEvent,
-    domain: TransitionDomain,
-    header_rebase: HeaderInsertionRebase,
-    no_change_effect: Option<TransitionEffect>,
-}
-
-/// Check replay identity, async ownership, and version freshness for an admitted request.
-///
-/// Returns `no_change_effect` when the event matches the last committed fingerprint or the
-/// header is already present; otherwise returns the event for further planning. Conflicting
-/// replay keys or stale versions fail.
-fn bind_replay_and_freshness(
-    engine: &HeaderChainEngine,
-    input: &TransitionInput,
-    snapshot_before_commit: &EngineSnapshot,
-    metadata: &crate::EngineMetadata,
-    request: AdmittedRequest,
-) -> Result<BoundRequest, TransitionFailure> {
-    let AdmittedRequest {
-        mut event,
-        expected_version,
-    } = request;
-    let domain = event.domain();
-    let header_rebase =
-        rebase_header_insertion(&mut event, snapshot_before_commit, engine.graph(), input)?;
-    if let Some(owner) = event.header_sync_owner() {
-        validate_header_sync_owner(owner, snapshot_before_commit)?;
-    }
-    if let Some(owner) = event.body_owner() {
-        validate_body_owner(owner, snapshot_before_commit)?;
-    }
-
-    let fingerprint = event.fingerprint();
-    if fingerprint.is_some() && metadata.last_transition == fingerprint {
-        return Ok(BoundRequest {
-            event,
-            domain,
-            header_rebase,
-            no_change_effect: Some(TransitionEffect::event()),
-        });
-    }
-    if metadata
-        .last_transition
-        .zip(fingerprint)
-        .is_some_and(|(previous, current)| previous.conflicts_with(current))
-    {
-        return Err(TransitionFailure::ConflictingReplay);
-    }
-    let has_async_authority = event.header_sync_owner().is_some() || event.body_owner().is_some();
-    if !has_async_authority {
-        let Some(expected_version) = expected_version else {
-            return Err(TransitionFailure::Stale {
-                current: snapshot_before_commit.state_version,
-            });
-        };
-        if expected_version != snapshot_before_commit.state_version {
-            return Err(TransitionFailure::Stale {
-                current: snapshot_before_commit.state_version,
-            });
-        }
-    }
-    Ok(BoundRequest {
-        event,
-        domain,
-        header_rebase,
-        no_change_effect: (header_rebase == HeaderInsertionRebase::AlreadyApplied)
-            .then_some(TransitionEffect::header_work_already_applied()),
-    })
-}
-
 /// Assemble the durable write set and graph delta from a fully settled projection.
 fn assemble_writes<'a>(
     engine: &'a HeaderChainEngine,
     snapshot_before_commit: EngineSnapshot,
-    old_selected: &'a [Frontier],
-    old_verified: &'a [Frontier],
+    old_selected: &'a [crate::Frontier],
+    old_verified: &'a [crate::Frontier],
     settled: settlement::SettledTransition<'a>,
     event: &crate::TransitionEvent,
     context: &TransitionContext<'_>,
